app/requirements.txt
cryptography==41.0.7
email-validator==2.1.0.post1
Flask==3.0.0
flask-admin==1.6.1
flask-jwt-extended==4.5.3
flask-login @ git+https://github.com/maxcountryman/flask-login@main
flask-mailman==1.0.0
Flask-Migrate==4.0.5
flask-restful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
ipython
psycopg2-binary==2.9.9  # sqlalchemy does not support psycopg[binary,pool] yet
PyJWT==2.8.0
sqlalchemy-utils==0.41.1
